% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetaSheet.r
\name{calculateAminoAcidFreq}
\alias{calculateAminoAcidFreq}
\title{calculateAminoAcidFreq}
\usage{
calculateAminoAcidFreq(sequences, codonTable = "standard")
}
\arguments{
\item{sequences}{DNAStringSet of \code{sequences}}

\item{codonTable}{select codon table}
}
\value{
aaFreq
}
\description{
Create aaFreq to store codon frequency.
}
\author{
Hebin Zhang
}
