% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIndexes.r
\name{calculateCAI}
\alias{calculateCAI}
\title{calculateCAI}
\usage{
calculateCAI(sequences, highExpressedGene, codonTable = "standard")
}
\arguments{
\item{sequences}{DNAStringSet of \code{sequences}}

\item{highExpressedGene}{highExpressedGene DNAStringSet of \code{highExpressedGene}}

\item{codonTable}{select codon table}
}
\value{
cai
}
\description{
calculate cai(codon adaptation index).
}
\author{
Hebin Zhang
}
