% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetaSheet.r
\name{calculateCodonFreq}
\alias{calculateCodonFreq}
\title{calculateCodonFreq}
\usage{
calculateCodonFreq(sequences)
}
\arguments{
\item{sequences}{DNAStringSet of \code{sequences}}
}
\value{
codonFreq
}
\description{
Create codonFreq to store codon frequency.
}
\author{
Hebin Zhang
}
