% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIndexes.r
\name{calculateRelAdapt}
\alias{calculateRelAdapt}
\title{calculateRelAdapt}
\usage{
calculateRelAdapt(highExpressedGene, codonTable = "standard")
}
\arguments{
\item{highExpressedGene}{highExpressedGene DNAStringSet of \code{highExpressedGene}}

\item{codonTable}{select codon table \code{codonTable}}
}
\value{
relAdapt
}
\description{
calculate relative adaptability.
}
\author{
Hebin Zhang
}
