% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIndexes.r
\name{calculateRSCU}
\alias{calculateRSCU}
\title{calculatecalculateRSCU}
\usage{
calculateRSCU(sequences, codonTable = "standard")
}
\arguments{
\item{sequences}{DNAStringSet of \code{sequences}}

\item{codonTable}{select codon table \code{codonTable}}
}
\value{
rscuSheet
}
\description{
calculate sequences RSCU.
}
\author{
Hebin Zhang
}
