% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{exampleSequence}
\alias{exampleSequence}
\title{an example sequence}
\format{
fasta.
}
\source{
Down from NCBI
}
\description{
an example sequence
}
\examples{
data(exampleSequence)
str(hiNorm)
}
