% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIndexes.r
\name{findOptimalCodonsBinomial}
\alias{findOptimalCodonsBinomial}
\title{findOptimalCodonsBinomial}
\usage{
findOptimalCodonsBinomial(highFile, lowFile)
}
\arguments{
\item{highFile}{Enter the fasta file of the highly expressed gene}

\item{lowFile}{Enter the fasta file of the lowly expressed gene}
}
\value{
binomialResults
}
\description{
Finding Optimal codons using binomial regression.
}
\author{
Hebin Zhang
}
