% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonTables.r
\name{getCodonTable}
\alias{getCodonTable}
\title{getCodonTable}
\usage{
getCodonTable(specie)
}
\arguments{
\item{specie}{select a specie \code{specie}}
}
\value{
codonTable
}
\description{
return a codon table
}
\author{
Hebin Zhang
}
