% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIndexes.r
\name{calculateFOP}
\alias{calculateFOP}
\title{calculateFOP}
\usage{
calculateFOP(codonFreq, optimalCodons)
}
\arguments{
\item{codonFreq}{codon frequency \code{codonFreq}}

\item{optimalCodons}{Enter optimal Codons \code{lowFile}}
}
\value{
fop
}
\description{
calculate FOP(frequency of optimal codons)
}
\author{
Hebin Zhang
}
